Handle issues more gracefully
When a vendor's registration in SAM is not "ACTIVE", the `certifications`
hash is `nil`, so check to make sure that there actually *are*
certifications.

Similarly, if there's only *one* NAICS code, the `find` method freaks out,
so we check the `small_biz_array` to see if it's actually an Array before
we run the `find` method.